+++ printcetnrum_web.client/src/app/app-routing.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomePageComponent } from './home-page/home-page.component';
-import { LoginPageComponent } from './login-page/login-page.component';
-import { RegisterPageComponent } from './register-page/register-page.component';
-import { EditPhotoPageComponent } from './edit-photo-page/edit-photo-page.component';
-import { UploadPhotoPageComponent } from './upload-photo-page/upload-photo-page.component';
-import { UsersListPageComponent } from './users-list-page/users-list-page.component';
+import { HomePageComponent } from './pages/home-page/home-page.component';
+import { LoginPageComponent } from './pages/login-page/login-page.component';
+import { RegisterPageComponent } from './pages/register-page/register-page.component';
+import { EditPhotoPageComponent } from './pages/edit-photo-page/edit-photo-page.component';
+import { UploadPhotoPageComponent } from './pages/upload-photo-page/upload-photo-page.component';
+import { UsersListPageComponent } from './pages/users-list-page/users-list-page.component';
 import { AuthGuard } from './guards/auth.guard';
 import { ResetPasswordPageComponent } from './user-accounts/reset-password-page/reset-password-page.component';
-import { UserFilesComponent } from './user-files/user-files.component';
+import { UserFilesComponent } from './pages/user-files/user-files.component';
 import { NewOrderComponent } from './orders/new-order/new-order.component';
 import { AllOrdersListComponent } from './orders/all-orders-list/all-orders-list.component';
 import { OrderDetailsComponent } from './orders/order-details/order-details.component';
