+++ printcetnrum_web.client/src/app/app.module.ts
@@ -10,18 +10,18 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { HomePageComponent } from './home-page/home-page.component';
-import { LoginPageComponent } from './login-page/login-page.component';
+import { HomePageComponent } from './pages/home-page/home-page.component';
+import { LoginPageComponent } from './pages/login-page/login-page.component';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { RegisterPageComponent } from './register-page/register-page.component';
-import { EditPhotoPageComponent } from './edit-photo-page/edit-photo-page.component';
-import { UploadPhotoPageComponent } from './upload-photo-page/upload-photo-page.component';
+import { RegisterPageComponent } from './pages/register-page/register-page.component';
+import { EditPhotoPageComponent } from './pages/edit-photo-page/edit-photo-page.component';
+import { UploadPhotoPageComponent } from './pages/upload-photo-page/upload-photo-page.component';
 import { FormsModule } from '@angular/forms';
-import { UsersListPageComponent } from './users-list-page/users-list-page.component';
+import { UsersListPageComponent } from './pages/users-list-page/users-list-page.component';
 import { TokenInterceptor } from './interceptors/token.interceptor';
 import { ResetPasswordPageComponent } from './user-accounts/reset-password-page/reset-password-page.component';
-import { UserFilesComponent } from './user-files/user-files.component';
+import { UserFilesComponent } from './pages/user-files/user-files.component';
 import { NewOrderComponent } from './orders/new-order/new-order.component';
 import { AllOrdersListComponent } from './orders/all-orders-list/all-orders-list.component';
 import { OrderDetailsComponent } from './orders/order-details/order-details.component';
