+++ printcetnrum_web.client/src/app/pages/edit-photo-page/edit-photo-page.component.css
