+++ printcetnrum_web.client/src/app/pages/upload-photo-page/upload-photo-page.component.html
