+++ printcetnrum_web.client/src/app/pages/upload-photo-page/upload-photo-page.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
-import { FileHandlerService } from '../services/file-handler.service';
+import { FileHandlerService } from '../../services/file-handler.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
-import { SnackBarUtil } from '../shared/snackbar-util';
+import { SnackBarUtil } from '../../shared/snackbar-util';
 
 @Component({
   selector: 'app-upload-photo-page',
